Fix to float area sizes

+++ input.h
@@ -4,7 +4,7 @@ int inputVerification(int input, int lower_boundry, int upper_boundry);
 int ha_to_m2_converter(float ha);
 void error_if_input_is_outside_range(int *checker);
 int get_input(int is_ellenberg_value, int lower_boundry, int upper_boundry);
-int ellenberg_value_condition (int input, int is_ellenberg_value, int lower_boundry, int upper_boundry);
+int ellenberg_value_condition (float input, int is_ellenberg_value, float lower_boundry, float upper_boundry);
 
 struct area read_input(void) {
   struct area area;
@@ -73,11 +73,10 @@ void error_if_input_is_outside_range(int *checker) {
   *checker += 1;
 }
 
-int ellenberg_value_condition(int input, int is_ellenberg_value, int lower_boundry, int upper_boundry) {
+int ellenberg_value_condition(float input, int is_ellenberg_value, float lower_boundry, float upper_boundry) {
   if (is_ellenberg_value == 1) {
     return !inputVerification(input, lower_boundry, upper_boundry);
   }
-  
+
   return input <= 0.0;
 }
-
